backend-fastapi/requirements.txt
# API Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Database
sqlalchemy>=2.0.36
aiosqlite>=0.20.0
greenlet>=3.0.0

# AI / LLM
openai>=1.50.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0

# Vector DB (local dev: Qdrant in-memory | UAT: Qdrant server or Azure AI Search)
qdrant-client>=1.9.0
langchain-qdrant>=0.1.0

# Document handling
python-multipart>=0.0.9
Pillow>=11.0.0
pypdf>=4.2.0
pymupdf>=1.24.0

# Validation & utilities
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.0
thefuzz>=0.22.0
Levenshtein>=0.26.0

# HTTP client
httpx>=0.27.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Kafka
kafka-python>=2.0.2
